+++ docs/conf.py
@@ -12,14 +12,17 @@ import sys
 
 # Find the path to the source files we want to to document, relative to the location of this file,
 # convert it to an absolute path.
-py_path = os.path.join(os.getcwd(), os.path.dirname(__file__), '../src')
+# Sphinx-multiversion checks out the repo at different ref points into tmp directory.
+# We're using SPHINX_MULTIVERSION_SOURCEDIR in order to access each such directory.
+# See https://github.com/Holzhaus/sphinx-multiversion/issues/42 for details.
+default_py_root = os.path.join(os.getcwd(), os.path.dirname(__file__))
+py_path = os.path.join(os.getenv("SPHINX_MULTIVERSION_SOURCEDIR", default=default_py_root), "../src")
 sys.path.insert(0, os.path.abspath(py_path))
-
 
 # -- Project information -----------------------------------------------------
 
 project = 'IQM client'
-copyright = '2021, IQM'
+copyright = '2021-2022, IQM'
 author = 'IQM client developers'
 
 # The short X.Y version.
@@ -32,7 +35,6 @@ except ImportError:
     pass
 else:
     release = version
-
 
 # -- General configuration ---------------------------------------------------
 
@@ -51,6 +53,7 @@ extensions = [
     'sphinx.ext.todo',
     'sphinx.ext.extlinks',
     'sphinx.ext.intersphinx',
+    "sphinx_multiversion",
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -112,6 +115,12 @@ html_theme_path = [sphinx_book_theme.get_html_theme_path()]
 # documentation.
 html_theme_options = {
     'logo_only': True,
+}
+
+html_sidebars = {
+    '**': [
+        'sidebar-logo.html', 'search-field.html', 'sbt-sidebar-nav.html', 'versioning.html'
+    ]
 }
 
 # A shorter title for the navigation bar.  Default is the same as html_title.
@@ -177,3 +186,13 @@ extlinks = {
 
 # List of all bibliography files used.
 #bibtex_bibfiles = ['references.bib']
+
+# -- Options for sphinx_multiversion --------------------------------------------------
+# Only include certain tags (i.e. all tags except for ones listed below)
+# (technically a whitelist, but we treat it as blacklist by using negative lookahead regex `?!`)
+smv_tag_whitelist = r'^(?!(1\.0|1\.1\b|1\.2|1\.3|1\.4|1\.5|1\.6|1\.7|1\.8|1\.9|2\.0|2\.1|4\.1)).*$'
+
+smv_branch_whitelist = "None"   # Do not include local branches in versions list
+smv_remote_whitelist = "None"   # Do not include remote branches in versions list
+smv_released_pattern = r'^refs/tags/.*$'  # Tags recognized as releases
+smv_outputdir_format = 'versions/{ref.name}'  # Store versioned docs in a subdirectory